added "run-tests" and "pre-commit"
"run-test" is a script, that performs all tests in the current work
tree. This should be executed during development.

"pre-commit" is a wrapper arount "run-test" such, that it can be
symlinked to serve as a git pre-commit hook. Doing so will ensure to run
all tests before commiting. The wrapper also stashes changes that will not
be commited and restores them afterwards.