return a histogram == 0 everywhere, if no data is available
this changes the default behaviour, if a histogram is requested, but the
MultiSpecies object contains no particles. Lets say you want to subtract
the charge density dumped with the
charge density calculated by postpic to compare both. In case there is a
dump without any particles this will now work (because a field object
with correct shape is returned), but failed before this
commit.