openPMD: Fix Reader of Scalar Records
For scalar openPMD `records` such as a temperature field,
a particle charge attribute, etc. the `record` itself is identical
to its `record component`.

This commit fixes the detection of `record` attributes specific
to fields (meshes) by checking either on the `record component`
(scalar records) and if the attribute is not found there at the
`parent` group (for vector records) level.