fixing map_coordinates
Scipy intoduced another paramter for padding, which apparently was zero
before. Fixing this.

The scipy change is:
https://github.com/scipy/scipy/commit/9c299da04b9c419786a56c8e657aaabeb35f9069

Alex noted, that on his implementation, using the normal
`map_coordinates` function was really slow. (I am guessing due to extra
memory allocation), which is why he called the C-Function directly and
did all the checks within postpic. So keep it as is for really fast
coordinate mapping :)

Fix #261